chore(mailchimp): typescript and move enforceSubscriptions to mailchimp module
